refactor(types): объявленный тип параметра вынесен из инференсера
Продолжение выноса знаний, не относящихся к выводу типов. Инференсер держал правила о
том, откуда берётся объявленный тип параметра: документирующий комментарий метода,
контракт платформенного события для обработчиков, одноимённый параметр метода из ссылки
`// См. …`.

Вынесено в `DeclaredParameterTypes`. Группа оказалась самой развязанной из оставшихся:
обратного вызова к выводу типов ей не нужно вовсе — она отвечает только про то, что
объявлено вне тела метода. Вместе с ней уехали индекс контрактов событий и поиск метода
модуля по имени: за пределами этой группы они не использовались.

Инференсер: 1527 строк → 1415, от 1650 на develop.
